backend/requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.2
Pillow>=10.0
requests>=2.31
python-dotenv>=1.0
openpyxl>=3.1
reportlab>=4.0
pytest>=8.0
pytest-django>=4.8
